updated file names, update docker files with new packages
